+++ rubocop-airbnb/lib/rubocop/cop/airbnb/const_assigned_in_wrong_file.rb
@@ -22,7 +22,7 @@ module RuboCop
       #   module Foo
       #     BAZ = 42
       #   end
-      class ConstAssignedInWrongFile < Cop
+      class ConstAssignedInWrongFile < Base
         include Inflections
         include RailsAutoloading