+++ rubocop-airbnb/lib/rubocop/cop/airbnb/simple_unless.rb
@@ -3,7 +3,7 @@ module RuboCop
     module Airbnb
       # Cop to tackle prevent unless statements with multiple conditions
       # https://github.com/airbnb/ruby#unless-with-multiple-conditions
-      class SimpleUnless < Cop
+      class SimpleUnless < Base
         MSG = 'Unless usage is okay when there is only one conditional'.freeze
 
         def_node_matcher :multiple_conditionals?, '(if ({and or :^} ...) ...)'