+++ rubocop-airbnb/spec/rubocop/cop/airbnb/mass_assignment_accessible_modifier_spec.rb
@@ -1,28 +1,23 @@
-describe RuboCop::Cop::Airbnb::MassAssignmentAccessibleModifier do
-  subject(:cop) { described_class.new }
-
+describe RuboCop::Cop::Airbnb::MassAssignmentAccessibleModifier, :config do
   it 'rejects when accessible= is called' do
-    source = [
-      'def some_method',
-      '  user = User.new',
-      '  user.accessible = [:first_name, :last_name]',
-      '  user.update_attributes(:first_name => "Walter", :last_name => "Udoing")',
-      'end',
-    ].join("\n")
-    inspect_source(source)
-    expect(cop.offenses.size).to eq(1)
+    expect_offense(<<~RUBY)
+      def some_method
+        user = User.new
+        user.accessible = [:first_name, :last_name]
+        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Do no override and objects mass assignment restrictions.
+        user.update_attributes(:first_name => "Walter", :last_name => "Udoing")
+      end
+    RUBY
   end
 
   it 'accepts when accessible= is not called' do
-    source = [
-      'def some_method',
-      '  user = User.new',
-      '  user.first_name = "Walter"',
-      '  user.last_name = "Udoing"',
-      '  user.save!',
-      'end',
-    ].join("\n")
-    inspect_source(source)
-    expect(cop.offenses).to be_empty
+    expect_no_offenses(<<~RUBY)
+      def some_method
+        user = User.new
+        user.first_name = "Walter"
+        user.last_name = "Udoing"
+        user.save!
+      end
+    RUBY
   end
 end