+++ rubocop-airbnb/spec/rubocop/cop/airbnb/risky_activerecord_invocation_spec.rb
@@ -1,54 +1,40 @@
-describe RuboCop::Cop::Airbnb::RiskyActiverecordInvocation do
-  subject(:cop) { described_class.new }
-
+describe RuboCop::Cop::Airbnb::RiskyActiverecordInvocation, :config do
   it "allows where statement that's a hash" do
-    source = [
-      'Users.where({:name => "Bob"})',
-    ].join("\n")
-
-    inspect_source(source)
-    expect(cop.offenses).to be_empty
+    expect_no_offenses(<<~RUBY)
+      Users.where({:name => "Bob"})
+    RUBY
   end
 
   it "allows where statement that's a flat string" do
-    source = 'Users.where("age = 24")'
-
-    inspect_source(source)
-    expect(cop.offenses).to be_empty
+    expect_no_offenses('Users.where("age = 24")')
   end
 
   it "allows a multiline where statement" do
-    source = "Users.where(\"age = 24 OR \" \\\n\"age = 25\")"
-
-    inspect_source(source)
-    expect(cop.offenses).to be_empty
+    expect_no_offenses("Users.where(\"age = 24 OR \" \\\n\"age = 25\")")
   end
 
   it "allows interpolation in subsequent arguments to where" do
-    source = 'Users.where("name like ?", "%#{name}%")'
-
-    inspect_source(source)
-    expect(cop.offenses).to be_empty
+    expect_no_offenses('Users.where("name like ?", "%#{name}%")')
   end
 
   it "disallows interpolation in where statements" do
-    source = 'Users.where("name = #{username}")'
-
-    inspect_source(source)
-    expect(cop.offenses.size).to eq(1)
+    expect_offense(<<~RUBY)
+      Users.where("name = \#{username}")
+      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Passing a string computed by interpolation or addition to an ActiveRecord method is likely to lead to SQL injection. [...]
+    RUBY
   end
 
   it "disallows addition in where statements" do
-    source = 'Users.where("name = " + username)'
-
-    inspect_source(source)
-    expect(cop.offenses.size).to eq(1)
+    expect_offense(<<~RUBY)
+      Users.where("name = " + username)
+      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Passing a string computed by interpolation or addition to an ActiveRecord method is likely to lead to SQL injection. [...]
+    RUBY
   end
 
   it "disallows interpolation in order statements" do
-    source = 'Users.where("age = 24").order("name #{sortorder}")'
-
-    inspect_source(source)
-    expect(cop.offenses.size).to eq(1)
+    expect_offense(<<~RUBY)
+      Users.where("age = 24").order("name \#{sortorder}")
+      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Passing a string computed by interpolation or addition to an ActiveRecord method is likely to lead to SQL injection. [...]
+    RUBY
   end
 end