+++ rubocop-airbnb/spec/rubocop/cop/airbnb/spec_constant_assignment_spec.rb
@@ -1,80 +1,64 @@
-describe RuboCop::Cop::Airbnb::SpecConstantAssignment do
-  subject(:cop) { described_class.new }
-
+describe RuboCop::Cop::Airbnb::SpecConstantAssignment, :config do
   it 'rejects constant definition inside of a describe block' do
-    source = [
-      'describe Someclass do',
-      '  CONSTANT = 5',
-      'end',
-    ].join("\n")
-
-    inspect_source(source)
-    expect(cop.offenses.size).to eq(1)
+    expect_offense(<<~RUBY)
+      describe Someclass do
+        CONSTANT = 5
+        ^^^^^^^^^^^^ Defining constants inside of specs can cause spurious behavior. [...]
+      end
+    RUBY
   end
 
   it 'allows constant defined inside of a module' do
-    source = [
-      'module Someclass',
-      '  CONSTANT = 5',
-      'end',
-    ].join("\n")
-
-    inspect_source(source)
-    expect(cop.offenses).to be_empty
+    expect_no_offenses(<<~RUBY)
+      module Someclass
+        CONSTANT = 5
+      end
+    RUBY
   end
 
-  it 'allows constant defined in global space' do
-    source = [
-      'CONSTANT = 5',
-    ].join("\n")
-
-    inspect_source(source)
-    expect(cop.offenses.size).to eq(1)
+  # TODO: The text of this statement did not match the test; which is correct
+  # behaviour?
+  it 'rejects constant defined in global space' do
+    expect_offense(<<~RUBY)
+      CONSTANT = 5
+      ^^^^^^^^^^^^ Defining constants inside of specs can cause spurious behavior. [...]
+    RUBY
   end
 
   it 'rejects constant assignment inside a before block' do
-    source = [
-      'describe Someclass do',
-      '  before { CONSTANT = 5 }',
-      'end',
-    ].join("\n")
-
-    inspect_source(source)
-    expect(cop.offenses.size).to eq(1)
+    expect_offense(<<~RUBY)
+      describe Someclass do
+        before { CONSTANT = 5 }
+                 ^^^^^^^^^^^^ Defining constants inside of specs can cause spurious behavior. [...]
+      end
+    RUBY
   end
 
   it 'rejects namespaced constant assignment inside a before block' do
-    source = [
-      'describe Someclass do',
-      '  before { MyModule::CONSTANT = 5 }',
-      'end',
-    ].join("\n")
-
-    inspect_source(source)
-    expect(cop.offenses.size).to eq(1)
+    expect_offense(<<~RUBY)
+      describe Someclass do
+        before { MyModule::CONSTANT = 5 }
+                 ^^^^^^^^^^^^^^^^^^^^^^ Defining constants inside of specs can cause spurious behavior. [...]
+      end
+    RUBY
   end
 
   it 'rejects constant assignment inside it block' do
-    source = [
-      'describe Someclass do',
-      '  it "tests stuff" do',
-      '    CONSTANT = 5',
-      '  end',
-      'end',
-    ].join("\n")
-
-    inspect_source(source)
-    expect(cop.offenses.size).to eq(1)
+    expect_offense(<<~RUBY)
+      describe Someclass do
+        it "tests stuff" do
+          CONSTANT = 5
+          ^^^^^^^^^^^^ Defining constants inside of specs can cause spurious behavior. [...]
+        end
+      end
+    RUBY
   end
 
   it 'allows let statements that do not assign constants' do
-    source = [
-      'describe Someclass do',
-      '  let(:constant) { 5 }',
-      'end',
-    ].join("\n")
-
-    inspect_source(source)
-    expect(cop.offenses).to be_empty
+    expect_no_offenses(<<~RUBY)
+      describe Someclass do
+        let(:constant) { 5 }
+      end
+    RUBY
   end
 end